pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "renoir-wikiart"
version = "3.0.0"
description = "A pedagogical tool for analyzing artist-specific works from WikiArt with computational color theory capabilities"
readme = "README.md"
authors = [
    {name = "Michail Semoglou", email = "m.semoglou@tongji.edu.cn"}
]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Artistic Software",
]
keywords = ["art-history", "dataset-analysis", "digital-humanities", "education", "computational-design", "color-theory"]
requires-python = ">=3.8"
dependencies = [
    "datasets>=2.0.0",
    "Pillow>=8.0.0",
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["renoir*"]
exclude = ["tests*", "examples*", "Research*", "test_env*", "Pierre-Auguste Renoir*", "jose*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "jupyter>=1.0.0",
]
visualization = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]

[project.urls]
Homepage = "https://github.com/MichailSemoglou/renoir"
Documentation = "https://github.com/MichailSemoglou/renoir#readme"
Repository = "https://github.com/MichailSemoglou/renoir"
"Bug Tracker" = "https://github.com/MichailSemoglou/renoir/issues"
